refactor: 최상위 lotto 패키지 정의
- model, view, controller 패키지들을 상위 lotto 패키지로 묶음
