--- src/main/kotlin/Application.kt
@@ -1,5 +1,0 @@
-import controller.LotteryController
-
-fun main() {
-    LotteryController().start()
-}
