+++ src/main/kotlin/lotto/Application.kt
@@ -1,0 +1,7 @@
+package lotto
+
+import lotto.controller.LotteryController
+
+fun main() {
+    LotteryController().start()
+}
