+++ src/main/kotlin/lotto/model/PurchaseInformation.kt
@@ -1,0 +1,6 @@
+package lotto.model
+
+data class PurchaseInformation(
+    val amount: Amount,
+    val manualLotteryCount: ManualLotteryCount,
+)
