+++ src/main/kotlin/lotto/model/Ticket.kt
@@ -1,0 +1,6 @@
+package lotto.model
+
+data class Ticket(
+    val userLotteries: List<Lottery>,
+    val purchaseInformation: PurchaseInformation,
+)
