+++ src/main/kotlin/lotto/model/generator/LotteriesGenerationStrategy.kt
@@ -1,0 +1,7 @@
+package lotto.model.generator
+
+import lotto.model.Lottery
+
+interface LotteriesGenerationStrategy {
+    fun issueLotteries(): List<Lottery>
+}
