+++ src/main/kotlin/lotto/model/generator/LotteryNumberGenerator.kt
@@ -1,0 +1,7 @@
+package lotto.model.generator
+
+import lotto.model.LotteryNumber
+
+interface LotteryNumberGenerator {
+    fun generateNumbers(): Set<LotteryNumber>
+}
