+++ src/main/kotlin/lotto/model/generator/ManualLotteriesGenerationStrategy.kt
@@ -1,0 +1,11 @@
+package lotto.model.generator
+
+import lotto.model.Lottery
+
+class ManualLotteriesGenerationStrategy(
+    private val manualLotteriesNumbers: List<List<String>>,
+) : LotteriesGenerationStrategy {
+    override fun issueLotteries(): List<Lottery> {
+        return manualLotteriesNumbers.map { Lottery.fromInput(it) }
+    }
+}
