--- src/main/kotlin/model/LotteryMachine.kt
@@ -1,23 +1,0 @@
-package model
-
-import model.generator.LotteriesGenerationStrategy
-import model.generator.ManualLotteriesGenerationStrategy
-
-object LotteryMachine {
-    fun issueLotteries(lotteriesGenerationStrategy: LotteriesGenerationStrategy): List<Lottery> {
-        return lotteriesGenerationStrategy.issueLotteries()
-    }
-
-    fun issueTicket(
-        manualLotteriesInput: List<List<String>>,
-        strategyForAuto: LotteriesGenerationStrategy,
-        purchaseInformation: PurchaseInformation,
-    ): Ticket {
-        val manualLotteries = issueLotteries(ManualLotteriesGenerationStrategy(manualLotteriesInput))
-        val autoLotteries = issueLotteries(strategyForAuto)
-
-        val userLotteries = manualLotteries + autoLotteries
-
-        return Ticket(userLotteries, purchaseInformation)
-    }
-}
