--- src/main/kotlin/model/PurchaseInformation.kt
@@ -1,6 +1,0 @@
-package model
-
-data class PurchaseInformation(
-    val amount: Amount,
-    val manualLotteryCount: ManualLotteryCount,
-)
