--- src/main/kotlin/model/Ticket.kt
@@ -1,6 +1,0 @@
-package model
-
-data class Ticket(
-    val userLotteries: List<Lottery>,
-    val purchaseInformation: PurchaseInformation,
-)
