--- src/main/kotlin/model/generator/LotteriesGenerationStrategy.kt
@@ -1,7 +1,0 @@
-package model.generator
-
-import model.Lottery
-
-interface LotteriesGenerationStrategy {
-    fun issueLotteries(): List<Lottery>
-}
