--- src/main/kotlin/model/generator/LotteryNumberGenerator.kt
@@ -1,7 +1,0 @@
-package model.generator
-
-import model.LotteryNumber
-
-interface LotteryNumberGenerator {
-    fun generateNumbers(): Set<LotteryNumber>
-}
