--- src/test/kotlin/model/generator/ExplicitLotteriesGenerationStrategy.kt
@@ -1,9 +1,0 @@
-package model.generator
-
-import model.Lottery
-
-class ExplicitLotteriesGenerationStrategy(
-    private val explicitLotteriesNumbers: List<Set<Int>>,
-) : LotteriesGenerationStrategy {
-    override fun issueLotteries(): List<Lottery> = explicitLotteriesNumbers.map { Lottery.fromSet(it) }
-}
